Renamed pulse wave breed to gaussian/gauss  listOfViewClasses -> singel DEFAULT_VIEW_CLASS_NAME in panel  listOfViewClasses -> its own file  made canvas width == window width adjustable  App and Panel pass thru width  made canvas height adjustable with size box  change default stepsPerIteratioin, lowPassFilter, wave type & freq  added panel.createdSpacePromise  added allowRunningOneCycle to alow removing it! sheesh  rename currentView to effectiveView  many more props in SquishView  renamed ResolutionDialog.openResolutionDialog to openResDialog  rename finalParams.viewClassName to mainViewClassName adjust fourier filter; some cleanup change meaning of lowPassFilter param (with SquishPanel) removed superlfours loadViewBuffer() removed dilution and 'lowPassFilter()'

+++ quantumEngine/schrodinger/Avatar.h
@@ -28,7 +28,7 @@ struct Avatar {
 
 	// params that the user can set
 	double dt;
-	double lowPassFilter;
+	int lowPassFilter;
 	int stepsPerIteration;
 
 	bool isIterating;
@@ -68,7 +68,7 @@ struct Avatar {
 	void visscherHalfStep(qWave *oldQWave, qWave *newQWave);  // obsolete
 
 	// kill high frequencies via FFTs
-	void fourierFilter(double lowPassFilter);
+	void fourierFilter(int lowPassFilter);
 
 	// set pleaseFFt from JS (only if in the middle of an iteration)
 	void askForFFT(void);
